refactor(api-types): DTO 및 codec 파일 분리
- User, UserProfile의 DTO 및 dateTimeCodec 별개 파일로 분리
- request, response 타입 정의용 schema를 담은 파일은 shared.ts -> helper.ts로 renaming
