+++ packages/api-types/src/users/current.ts
@@ -1,10 +1,10 @@
-import z from "zod";
-import { responseDtoSchema } from "../shared";
-import { UserDtoSchema } from "../auth/login";
-
 /*
  GET /users/current
  */
+
+import z from "zod";
+import { responseDtoSchema } from "../helpers";
+import { UserDtoSchema } from "../common/user-dto";
 
 export const GetCurrentUserResponseDtoSchema = responseDtoSchema(
   UserDtoSchema,
